refactor: use CAMLnoret instead of legacy macros
Previously we used `CAMLnoreturn_start` and `CAMLnoreturn_end`. These
are legacy macros avaliable for compatibility reasons. They have been
subsumed by the CAMLnoret macro (not to be confused with memory.h's
CAMLnoreturn, which is used in function bodies), to be placed at the
very start of a declaration or definition.