feat!: new api (#1)
Overhauls the API.

## User-facing updates

This updates the user-facing API relatively substantially. Most important is (1) the addition of a separate module for each of interpreted and JIT matching (with the option to also have one for DFAs, but not yet implemented) and (2) changes to matching functions.

Migration summary:
- Create a compiled regular expression given a pattern: regexp -> compile
- Check if a regex matches a string: pmatch -> is_match
- Match a regex without capture groups: N/A -> find
- Repeatedly match a regex without capture groups: N/A -> find_iter
- Match a regex with capture groups: exec -> captures
- Repeatedly match a regex with capture groups: exec_all -> captures_iter
- Perform templated replacement: replace, others -> N/A---in Semgrep see Pcre2_
- Extract substrings: extract, others -> Use find_iter and match functions

Support for callouts is removed (at least for now). Functions no longer take an optional regex or string, but require a pre-compiled regex. This should help resolve ambiguity and encourage re-using the same regex, rather than the mess that previously existed with a "default" regex.

We also replace almost all use of exceptions for user-facing errors with explicit error types and results.

## Internals

This features a re-write of the PCRE2 C bindings to help clean up the code a bit by both removing features we don't care about and documenting it more clearly. See `pcre2_stubs.c` for more. Code is best reviewed as just the new; the diff isn't super useful imo.

We introduce a new `Bindings` module to more clearly separate the main OCaml implementation from functions which just correspond to the C stubs directly.