fix: correctly report non-matching capture groups
When a capture group is not matched, but is between two that did, we
currently incorrectly give the range as `Some [-1, -1]`. This should
instead be `None`--i.e., no match. This was due to the fact that numbering
of capture groups is based on their order in the pattern text, but they
could be alternative or otherwise optional. For instance:

    (a)(b)?(c)
    (a)(?:(b)|(c))

In both above cases, the capture group `(b)` is number 2, but may not be
matched, either since it is optional or that case need not be taken.

To fix, we need to explicitly check that even if we're in bounds that
the capture group actually produced a match. See also `PCRE2_UNSET`,
`pcre2_match(3)`.

Co-authored-by: Matthew McQuaid <47466291+mmcqd@users.noreply.github.com>